Refactoring ChartPage
