+++ lib/ui/chart/chart_page.dart
@@ -1,23 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:money_writer_app/provider/transactions_provider.dart';
-import 'package:money_writer_app/ui/chart/pie_chart_transactions.dart';
 import 'package:money_writer_app/utils/result_state.dart';
-import 'package:month_picker_dialog/month_picker_dialog.dart';
+import 'package:money_writer_app/widgets/chartpage_month_picker.dart';
+import 'package:money_writer_app/widgets/chartpage_piechart_transactions.dart';
 import 'package:provider/provider.dart';
 
 class ChartPage extends StatelessWidget {
   static const routeName = '/chart_page';
 
   const ChartPage({Key? key}) : super(key: key);
-
-  // DateTime? selectedDate;
-
-  // @override
-  // void initState() {
-  //   super.initState();
-  //   selectedDate = DateTime.now();
-  // }
 
   @override
   Widget build(BuildContext context) {
@@ -28,7 +20,7 @@ class ChartPage extends StatelessWidget {
       body: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          MonthPickerChart(),
+          ChartMonthPicker(),
           Expanded(
             child: Consumer<TransactionsProvider>(
               builder: (context, provider, child) {
@@ -107,89 +99,5 @@ class ChartPage extends StatelessWidget {
         ],
       ),
     );
-  }
-}
-
-class MonthPickerChart extends StatefulWidget {
-  const MonthPickerChart({Key? key}) : super(key: key);
-
-  @override
-  _MonthPickerChartState createState() => _MonthPickerChartState();
-}
-
-class _MonthPickerChartState extends State<MonthPickerChart> {
-  DateTime? selectedDate;
-
-  @override
-  void initState() {
-    super.initState();
-    selectedDate = DateTime.now();
-    Provider.of<TransactionsProvider>(context, listen: false)
-        .getTotalInMonth(selectedDate!.month, selectedDate!.year);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Consumer<TransactionsProvider>(builder: (context, provider, child) {
-      return Material(
-        child: Container(
-          padding: const EdgeInsets.all(8.0),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              IconButton(
-                icon: const Icon(Icons.chevron_left),
-                onPressed: () {
-                  setState(() {
-                    selectedDate = DateTime(
-                      selectedDate!.year,
-                      selectedDate!.month - 1,
-                    );
-                  });
-                  getData(provider);
-                },
-              ),
-              TextButton(
-                child: Text(
-                  DateFormat("MMMM yyyy", "id_ID").format(selectedDate!),
-                  // 'Month: ${selectedDate?.month} - ${selectedDate?.year}',
-                ),
-                onPressed: () {
-                  showMonthPicker(
-                    context: context,
-                    initialDate: DateTime.now(),
-                    firstDate: DateTime(DateTime.now().year - 10, 5),
-                    lastDate: DateTime(DateTime.now().year + 1, 9),
-                  ).then((date) {
-                    if (date != null) {
-                      setState(() {
-                        selectedDate = date;
-                      });
-                      getData(provider);
-                    }
-                  });
-                },
-              ),
-              IconButton(
-                icon: const Icon(Icons.chevron_right),
-                onPressed: () {
-                  setState(() {
-                    selectedDate = DateTime(
-                      selectedDate!.year,
-                      selectedDate!.month + 1,
-                    );
-                  });
-                  getData(provider);
-                },
-              ),
-            ],
-          ),
-        ),
-      );
-    });
-  }
-
-  void getData(TransactionsProvider provider) {
-    provider.getTotalInMonth(selectedDate!.month, selectedDate!.year);
   }
 }
